Keys to choose Heroes and syling of it are done. #2
There are keys to choose each hero now. Players can activate the the hero they wanted.

When they click the key, it activates the chosen hero's opitons and changing the style
of the hero.

When Hero chosen, the image of the Hero also changes. But it does not change back.

The coding not the best for these properties to make. But this is it that I can do now.
To make it better I need to learn new tricks. Thats why I am finishing this project with this commit.