The Project Start
The project is done by watching video from javascript30.com.
It is in the folder js30-2. I have made a couple of changes.

The new Project has been created but I have not much done for the new one.
It is on the Process. I will try to understand how we give these clock looking to
our divs.