project 12 is done. We have created a secret code and when user click this code we make something special in the page.
It is about detecting the key and dong something. It is almost the same with first project.
In this one we are detecting a group of keys and to do this using some array properties.
Because it is so similar I am not doing a new project with this lesson. And I am moving to the 13th video.