Project 22 is done
Project 22 is finished. It is a nice property. I am gonna make a clone of it.

Project 19, 20 and 21 is not working in my computer.
There is a problem when I try to reach my media such as mikrophone and video.
I will not froce to make these project. I am just moving on with the next ones.