Project js30-22-yearmfew is done.
It is a simple article. It shows the meaning of the unknown words when user comes on it with the mouse.

There is a dictionary.js file to keep the meaning of the words. There is a data variable for it.

I am not changing the style for now. It can be improved when this project used for a meaningful article.